refactor(locker): separate each packages
